Switch to NextJS
Switching from managing static HTML pages
to a static website generated from a NextJS
codebase. This allows the developer to use
React JS Framework and Typescript to build
code and render the layouts and then
generate a static HTML page from it.

This makes for easier to maintain code and
something that should be simpler.

Also paves the way for some data to be
dynamically fetched and parsed at build
time.