Restructure to new schema to support cups
Adds the idea of a Squad, a layer above teams, that belongs to the club. A squad plays as a 'team' in one or more cups and a league.

The API has been stripped out and now provides endpoints to fetch league games/teams and jpl cup games and teams